+++ src/sensors/SpectralModule.h
@@ -2,17 +2,17 @@
 
 #include "DFRobot_AS7341.h"
 struct spectralChannels {
-	uint16_t f1 = 0;
-	uint16_t f2 = 0;
-	uint16_t f3 = 0;
-	uint16_t f4 = 0;
-	uint16_t f5 = 0;
-	uint16_t f6 = 0;
-	uint16_t f7 = 0;
-	uint16_t f8 = 0;
-	uint16_t clear = 0;
-	uint16_t nir = 0;
-	uint16_t maxIntensity = 0;
+	int f1 = 0;
+	int f2 = 0;
+	int f3 = 0;
+	int f4 = 0;
+	int f5 = 0;
+	int f6 = 0;
+	int f7 = 0;
+	int f8 = 0;
+	int clear = 0;
+	int nir = 0;
+	int maxIntensity = 0;
 	int largestCh;
 	String strCh;
 	int fCount[8];
